4.0: tasks feature is deprecated and removed in ArangoDB v4.0

+++ tests/conftest.py
@@ -294,18 +294,19 @@ async def teardown():
             verify=False,
         )
 
-        # Remove all tasks
-        test_tasks = [
-            task
-            for task in await sys_db.tasks()
-            if task["name"].startswith("test_task")
-        ]
-        await asyncio.gather(
-            *(
-                sys_db.delete_task(task["id"], ignore_missing=True)
-                for task in test_tasks
+        if global_data.db_version < version.parse("4.0.0"):
+            # Remove all tasks
+            test_tasks = [
+                task
+                for task in await sys_db.tasks()
+                if task["name"].startswith("test_task")
+            ]
+            await asyncio.gather(
+                *(
+                    sys_db.delete_task(task["id"], ignore_missing=True)
+                    for task in test_tasks
+                )
             )
-        )
 
         # Remove all test users.
         tst_users = [
